Sun, Apr 7, 2024, 4:12 PM -07:00

+++ wk4/pset/recover/recover.c
@@ -43,7 +43,7 @@ int main(int argc, char *argv[])
             files++
             Open output file with name (filename)
             Copy data from old to new file -->
-            int data[photolength][512];
+            ulong data[photolength][512];
             fread(&data, )
             fwrite()
         }
